% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crmfunc.R
\docType{package}
\name{crmfunc}
\alias{crmfunc}
\alias{crmfunc-package}
\title{\code{crmfunc} package}
\description{
An R Package for Handling CRM Data
}
\details{
This package aims to assist R users who are dealing with customer data 
(names, addresses, phone numbers, etc.) whether or not it comes from a CRM system.

Additional material can be found in the 
\href{https://github.com/reportmort/crmfunc}{README} on GitHub
}
